added check to make save_to_model() safe

+++ django_walletpass/models.py
@@ -208,6 +208,8 @@ class PassBuilder:
         else:
             filename = f"{uuid.uuid1()}.pkpass"
 
+        if self.builded_pass_content is None:
+            raise ValueError(_("Cannot not save to model: builded_pass_content is None."))
         content = WalletpassContentFile(self.builded_pass_content)
         instance.data.delete()
         instance.data.save(filename, content)
